TFmini driver: improve to support various TFmini models and basic autosetup for TFMini-plus
A new "tfmini command -c XXXXXX" is introduced to help device configuration
directly from the flight controler. TFMini and TFMini-plus are both supported.

The parser has been improved to support receiving confirmation from the device
after a command has been written.

The parameter SENS_EN_TFMINI is implemented again to let user define the TFMini
model that will be used. This helps to validate commands and validate command's answers.

SENS_EN_TFMINI is also required because TFMini models don't offer the same capabilities
for FOV and MIN-DISTANCE.

The parser has been improved to provide signal quality for TFMini-plus.
The same should be done in the future for TFMini model.

The driver has been improved to autoconfigure the mode for TFMini-plus upon starting.
The same should be done in the future for TFMini model.

Signed-off-by: Aymeric Moizard <amoizard@spores.eu>